requirements.txt
requests
python-dotenv
serpapi
colorama
replicate
tqdm
concurrent-log-handler
urllib3